requirements.txt
# Hardware Control
RPi.GPIO>=0.7.1
smbus2>=0.4.1
adafruit-circuitpython-sht31d>=2.3.0
board>=1.0
digitalio>=3.3.0

# Web Interface
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
eventlet>=0.33.0

# Data Processing
numpy>=1.21.0
pandas>=1.5.0

# Scheduling and Time
schedule>=1.2.0
python-crontab>=2.7.0

# Configuration and Logging
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: For development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0